move boost manager integration tests to single file
