+++ boost_manager/tests/integrations/activator_tests.rs
@@ -1,4 +1,3 @@
-mod common;
 use boost_manager::{activator, db, OnChainStatus};
 use chrono::{DateTime, Duration as ChronoDuration, Duration, Timelike, Utc};
 use mobile_config::boosted_hex_info::{BoostedHex, BoostedHexInfo, BoostedHexes};
