+++ boost_manager/tests/integrations/common/mod.rs
@@ -1,8 +1,43 @@
+use async_trait::async_trait;
+use chrono::DateTime;
+use chrono::Utc;
 use file_store::file_sink::{FileSinkClient, Message as SinkMessage};
+use futures_util::{stream, StreamExt as FuturesStreamExt};
 use helium_proto::BoostedHexInfoV1 as BoostedHexInfoProto;
 use helium_proto::BoostedHexUpdateV1 as BoostedHexUpdateProto;
 use helium_proto::Message;
+use mobile_config::boosted_hex_info::BoostedHexInfo;
+use mobile_config::boosted_hex_info::BoostedHexInfoStream;
+use mobile_config::client::hex_boosting_client::HexBoostingInfoResolver;
+use mobile_config::client::ClientError;
 use tokio::{sync::mpsc::error::TryRecvError, time::timeout};
+
+#[derive(Debug, Clone)]
+pub struct MockHexBoostingClient {
+    pub boosted_hexes: Vec<BoostedHexInfo>,
+}
+
+impl MockHexBoostingClient {
+    pub fn new(boosted_hexes: Vec<BoostedHexInfo>) -> Self {
+        Self { boosted_hexes }
+    }
+}
+
+#[async_trait]
+impl HexBoostingInfoResolver for MockHexBoostingClient {
+    type Error = ClientError;
+
+    async fn stream_boosted_hexes_info(&mut self) -> Result<BoostedHexInfoStream, ClientError> {
+        Ok(stream::iter(self.boosted_hexes.clone()).boxed())
+    }
+
+    async fn stream_modified_boosted_hexes_info(
+        &mut self,
+        _timestamp: DateTime<Utc>,
+    ) -> Result<BoostedHexInfoStream, ClientError> {
+        Ok(stream::iter(self.boosted_hexes.clone()).boxed())
+    }
+}
 
 pub struct MockFileSinkReceiver {
     pub receiver: tokio::sync::mpsc::Receiver<SinkMessage>,
