+++ boost_manager/tests/integrations/main.rs
@@ -1,0 +1,6 @@
+mod common;
+
+mod activator_tests;
+mod purger_tests;
+mod updater_tests;
+mod watcher_tests;
